finished/requirements.txt
requests
Flask
psycopg[binary]
psycopg2-binary
python-dotenv
gunicorn
